Chatbot Tutorial Backend Fix
- like Planner, I moved chatbot tutorial stuff to its file and it updates backend to not repeatedly invoke tutorial on just localStorage. It reads backend to see if they've already done it.
